+++ lib/src/util/url_launcher_utils.dart
@@ -1,3 +1,5 @@
+import 'package:base_lib_pub/base_lib_pub.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
 /// Name: url_launcher_utils.dart
@@ -7,7 +9,7 @@ import 'package:url_launcher/url_launcher_string.dart';
 
 /// 通过url启动第三方app
 /// platform配置：
-/// android：
+/// android：manifest
 // <!-- Provide required visibility configuration for API level 30 and above -->
 // <queries>
 //   <!-- If your app checks for SMS support -->
@@ -20,29 +22,67 @@ import 'package:url_launcher/url_launcher_string.dart';
 //     <action android:name="android.intent.action.VIEW" />
 //     <data android:scheme="tel" />
 //   </intent>
+//   <!-- If your app checks for mailto support -->
+//   <intent>
+//     <action android:name="android.intent.action.VIEW" />
+//     <data android:scheme="mailto" />
+//   </intent>
 // </queries>
-/// iOS：
+/// iOS：info.plist
 // <key>LSApplicationQueriesSchemes</key>
 // <array>
 //   <string>sms</string>
 //   <string>tel</string>
+//   <string>mailto</string>
 // </array>
 
-extension StringUrlExt on String {
-  /// 是否是schemeUrl，需要通过url_launcher打开第三方app
-  bool get isSchemeUrl => !startsWith('http://') && !startsWith('https://');
+/// 通过url_launcher启动，如果无法启动，返回false
+Future<bool> launchMyUrl(
+  String url, {
+  LaunchMode mode = LaunchMode.externalApplication,
+}) async {
+  if (await canLaunchUrlString(url)) {
+    return launchUrlString(url, mode: mode);
+  }
 
-  /// 通过url_launcher启动，如果无法启动，返回false
-  Future<bool> launch() async {
-    if (await canLaunchUrlString(this)) {
-      return launchUrlString(this, mode: LaunchMode.externalApplication);
-    }
+  return false;
+}
 
+/// 发送邮件
+Future<bool> mailto({
+  required String target, // 收件人邮箱
+  String? cc, // 抄送
+  String? bcc, // 密送
+  String? title, // 标题(部分邮箱可能不支持中文标题或乱码)
+  String? body, // 内容
+}) async {
+  if (target.isEmptyOrNull()) {
     return false;
   }
-
-  /// 在浏览器中打开
-  Future<bool> launchInBrowser() {
-    return launchUrlString(this, mode: LaunchMode.externalApplication);
+  Map<String, String> params = {};
+  if (!isEmptyOrNull(cc)) {
+    params['cc'] = cc!;
   }
+  if (!isEmptyOrNull(bcc)) {
+    params['bcc'] = bcc!;
+  }
+  if (!isEmptyOrNull(title)) {
+    params['subject'] = title!.encodeUrl();
+  }
+  if (!isEmptyOrNull(body)) {
+    params['body'] = body!.encodeUrl();
+  }
+  return launchUrl(Uri(
+    scheme: 'mailto',
+    path: target,
+    query: params.entries.map((MapEntry<String, String> e) => '${e.key}=${e.value}').join('&'),
+  )..toString().logI(tag: 'url_launcher_utils'));
+}
+
+/// 调用拨号页面
+Future<bool> tel(String phone) async {
+  if (phone.isEmptyOrNull()) {
+    return false;
+  }
+  return 'tel:$phone'.launch();
 }